[UPD] cg: format and refactor code
