+++ code_generator/models/code_generator_module.py
@@ -1,23 +1,12 @@
-# -*- coding: utf-8 -*-
-
-import base64
-
+from odoo import models, fields, api, modules, tools
 import logging
+import base64
 import lxml
 import os
-from docutils.core import publish_string
-from odoo import models, fields, api, modules, tools
 from odoo.addons.base.models.ir_module import MyWriter
+from docutils.core import publish_string
 
 _logger = logging.getLogger(__name__)
-
-
-class Module(models.Model):
-    _inherit = "ir.module.module"
-
-    header_manifest = fields.Text(
-        string="Header", help="Header comment in __manifest__.py file."
-    )
 
 
 class CodeGeneratorModule(models.Model):
@@ -376,362 +365,3 @@ class CodeGeneratorModule(models.Model):
             o2m_models.mapped("view_ids").unlink()
             o2m_models.unlink()  # I need to delete the created tables
         return super(CodeGeneratorModule, self).unlink()
-
-
-class CodeGeneratorModuleExternalDependency(models.Model):
-    _name = "code.generator.module.external.dependency"
-    _description = "Code Generator Module External Dependency"
-
-    module_id = fields.Many2one(
-        "code.generator.module", "Module", ondelete="cascade"
-    )
-
-    depend = fields.Char(string="Dependency name")
-
-    application_type = fields.Selection(
-        selection=[("python", "python"), ("bin", "bin")],
-        string="Application Type",
-        default="python",
-    )
-
-    # TODO this is wrong, an hack because ir_module_module != code_generator_module
-    is_template = fields.Boolean(
-        string="Is template", help="Will be affect template module."
-    )
-
-
-class CodeGeneratorModuleDependency(models.Model):
-    _inherit = "ir.module.module.dependency"
-    _name = "code.generator.module.dependency"
-    _description = "Code Generator Module Dependency"
-
-    module_id = fields.Many2one(
-        "code.generator.module", "Module", ondelete="cascade"
-    )
-
-    depend_id = fields.Many2one("ir.module.module", "Dependency", compute=None)
-
-
-class CodeGeneratorModuleTemplateDependency(models.Model):
-    _inherit = "ir.module.module.dependency"
-    _name = "code.generator.module.template.dependency"
-    _description = (
-        "Code Generator Module Template Dependency, set by"
-        " code_generator_template"
-    )
-
-    module_id = fields.Many2one(
-        "code.generator.module", "Module", ondelete="cascade"
-    )
-
-    depend_id = fields.Many2one("ir.module.module", "Dependency", compute=None)
-
-
-class CodeGeneratorActWindow(models.Model):
-    _name = "code.generator.act_window"
-    _description = "Code Generator Act Window"
-
-    name = fields.Char(string="name")
-
-    # TODO use ir.model.data instead if id_name
-    id_name = fields.Char(
-        string="Action id", help="Specify id name of this action window."
-    )
-
-    code_generator_id = fields.Many2one(
-        comodel_name="code.generator.module",
-        string="Code Generator",
-        required=True,
-        ondelete="cascade",
-    )
-
-
-class CodeGeneratorMenu(models.Model):
-    _name = "code.generator.menu"
-    _description = "Code Generator Menu"
-
-    # TODO missing groups_id and active and web_icon or web_icon_data
-    code_generator_id = fields.Many2one(
-        comodel_name="code.generator.module",
-        string="Code Generator",
-        required=True,
-        ondelete="cascade",
-    )
-
-    # TODO use ir.model.data instead if id_name
-    id_name = fields.Char(
-        string="Menu id", help="Specify id name of this menu."
-    )
-
-    # TODO use ir.model.data instead if parent_id_name
-    parent_id_name = fields.Char(
-        string="Menu parent id",
-        help="Specify id name of parent menu, optional.",
-    )
-
-    sequence = fields.Integer(string="Sequence", default=10)
-
-    m2o_act_window = fields.Many2one(
-        comodel_name="code.generator.act_window",
-        string="Action Windows",
-        help="Act window to open when click on this menu.",
-    )
-
-
-class CodeGeneratorView(models.Model):
-    _name = "code.generator.view"
-    _description = "Code Generator View"
-
-    code_generator_id = fields.Many2one(
-        comodel_name="code.generator.module",
-        string="Code Generator",
-        required=True,
-        ondelete="cascade",
-    )
-
-    # TODO use ir.model.data instead if id_name
-    id_name = fields.Char(
-        string="View id", help="Specify id name of this view."
-    )
-
-    view_type = fields.Selection(
-        [
-            ("tree", "Tree"),
-            ("form", "Form"),
-            ("pivot", "Pivot"),
-            ("graph", "Graph"),
-            ("search", "Search"),  # Special
-            ("kanban", "Kanban"),
-            ("timeline", "Timeline"),
-            ("activity", "Activity"),
-            ("calendar", "Calendar"),
-            ("diagram", "Diagram"),
-        ],
-        default="form",
-        help="Choose view type to generate.",
-    )
-
-    view_name = fields.Char(string="View name")
-
-    m2o_model = fields.Many2one(
-        comodel_name="ir.model",
-        string="Code generator Model",
-        help="Model related with this report",
-    )
-
-    view_item_ids = fields.Many2many(
-        comodel_name="code.generator.view.item",
-        string="View item",
-        help="Item view to add in this view.",
-    )
-
-    has_body_sheet = fields.Boolean(
-        string="Sheet format",
-        help="Use sheet presentation for body of form view.",
-    )
-
-
-class CodeGeneratorViewItem(models.Model):
-    _name = "code.generator.view.item"
-    _description = "Code Generator View Item"
-
-    action_name = fields.Char(string="Action name")
-
-    sequence = fields.Integer(string="Sequence", default=1)
-
-    # TODO create HTML for specific label
-    label = fields.Char(string="Label")
-
-    item_type = fields.Selection(
-        [
-            ("field", "Field"),
-            ("button", "Button"),
-            ("html", "HTML"),
-            ("filter", "Filter"),
-            ("div", "Division"),
-            ("group", "Group"),
-            ("templates", "Templates"),
-            ("t", "T"),
-            ("ul", "UL"),
-            ("li", "LI"),
-            ("i", "I"),
-            ("strong", "Strong"),
-        ],
-        default="field",
-        help="Choose item type to generate.",
-    )
-
-    button_type = fields.Selection(
-        [
-            ("", ""),  # Default
-            ("btn-default", "Default"),  # Default
-            ("btn-primary", "Primary"),
-            ("btn-secondary", "Secondary"),  # Default
-            ("btn-link", "Link"),  # URL
-            ("btn-success", "Success"),  # Green
-            ("btn-warning", "Warning"),  # Yellow
-            ("btn-danger", "Danger"),  # Red
-            ("oe_highlight", "Highlight"),  # Primary
-            ("oe_stat_button", "Statistic"),  # Default
-        ],
-        default="",
-        help="Choose item type to generate.",
-    )
-
-    background_type = fields.Selection(
-        [
-            ("", ""),  # Default
-            ("bg-success", "Success"),  # Default
-            # ("bg-success-light", "Success light"),
-            ("bg-success-full", "Success full"),
-            ("bg-warning", "Warning"),
-            # ("bg-warning-light", "Warning light"),
-            ("bg-warning-full", "Warning full"),
-            ("bg-info", "Info"),
-            # ("bg-info-light", "Info light"),
-            ("bg-info-full", "Info full"),
-            ("bg-danger", "Danger"),
-            # ("bg-danger-light", "Danger light"),
-            ("bg-danger-full", "Danger full"),
-            ("bg-light", "Light"),
-            ("bg-dark", "Dark"),
-        ],
-        default="",
-        help="Choose background color of HTML.",
-    )
-
-    section_type = fields.Selection(
-        [
-            ("header", "Header"),
-            ("title", "Title"),
-            ("body", "Body"),
-            ("footer", "Footer"),
-        ],
-        default="body",
-        help="Choose item type to generate.",
-    )
-
-    colspan = fields.Integer(
-        string="Colspan",
-        default=1,
-        help="Use this to fill more column, check HTML table.",
-    )
-
-    placeholder = fields.Char(string="Placeholder")
-
-    password = fields.Boolean(string="Password", help="Hide character.")
-
-    icon = fields.Char(
-        string="Icon",
-        help="Example fa-television. Only supported with button.",
-    )
-
-    attrs = fields.Char(
-        string="Attributes",
-        help="Specific condition, search attrs for more information.",
-    )
-
-    is_required = fields.Boolean(string="Required")
-
-    is_invisible = fields.Boolean(string="Invisible")
-
-    is_readonly = fields.Boolean(string="Readonly")
-
-    is_help = fields.Boolean(string="Help")
-
-    has_label = fields.Boolean(string="Labeled", help="Label for title.")
-
-    parent_id = fields.Many2one(comodel_name="code.generator.view.item")
-
-    child_id = fields.One2many(
-        comodel_name="code.generator.view.item",
-        inverse_name="parent_id",
-    )
-
-    edit_only = fields.Boolean(string="Edit only")
-
-
-class CodeGeneratorPyClass(models.Model):
-    _name = "code.generator.pyclass"
-    _description = "Code Generator Python Class"
-
-    name = fields.Char(string="Class name", help="Class name", required=True)
-
-    module = fields.Char(string="Class path", help="Class path")
-
-
-class CodeGeneratorCodeImport(models.Model):
-    _name = "code.generator.model.code.import"
-    _description = "Header code to display in model"
-
-    name = fields.Char(string="Import name", help="import name")
-
-    sequence = fields.Integer(
-        string="Sequence", help="Order of sequence code."
-    )
-
-    code = fields.Text(
-        string="code", help="Code of import header of python file"
-    )
-
-    m2o_module = fields.Many2one(
-        "code.generator.module",
-        string="Module",
-        help="Module",
-        ondelete="cascade",
-    )
-
-    m2o_model = fields.Many2one(
-        "ir.model", string="Model", help="Model", ondelete="cascade"
-    )
-
-    is_templated = fields.Boolean(
-        string="Templated", help="Code for code generator from template."
-    )
-
-
-class CodeGeneratorCode(models.Model):
-    _name = "code.generator.model.code"
-    _description = "Code to display in model"
-
-    name = fields.Char(string="Method name", help="Method name", required=True)
-
-    sequence = fields.Integer(
-        string="Sequence", help="Order of sequence code."
-    )
-
-    code = fields.Text(
-        string="Code of pre_init_hook",
-        default="""
-return""",
-    )
-
-    decorator = fields.Char(
-        string="Decorator",
-        help="Like @api.model. Use ; for multiple decorator.",
-    )
-
-    param = fields.Char(string="Param", help="Like : name,color")
-
-    returns = fields.Char(
-        string="Return type", help="Annotation to return type value."
-    )
-
-    m2o_module = fields.Many2one(
-        "code.generator.module",
-        string="Module",
-        help="Module",
-        ondelete="cascade",
-    )
-
-    m2o_model = fields.Many2one(
-        "ir.model", string="Model", help="Model", ondelete="cascade"
-    )
-
-    is_wip = fields.Boolean(
-        string="Work in progress", help="Temporary function to be fill later."
-    )
-
-    is_templated = fields.Boolean(
-        string="Templated", help="Code for code generator from template."
-    )
