+++ code_generator/models/ir_model.py
@@ -1,14 +1,9 @@
-# -*- coding: utf-8 -*-
-
 import importlib
 import logging
-import re
 
 from odoo import models, fields, api, _, tools
-from odoo.addons.base.models.ir_model import SAFE_EVAL_BASE
-from odoo.exceptions import ValidationError, UserError, MissingError
-from odoo.tools.safe_eval import safe_eval
-from psycopg2._psycopg import ProgrammingError
+
+from odoo.exceptions import ValidationError
 from odoo.models import MAGIC_COLUMNS
 
 _logger = logging.getLogger(__name__)
@@ -20,227 +15,6 @@ MAGIC_FIELDS = MAGIC_COLUMNS + [
     "access_token",
     "access_warning",
 ]
-CONSCREATEUNABLE = _("Unable to create the constraint.")
-CONSDELETECREATEUNABLE = _(
-    "Since you modify the sql constraint definition we must delete it and"
-    " create a new one, and we were unable to do it."
-)
-CONSMODIFYUNABLE = _("Unable to modify the constraint.")
-CONSDELETEUNABLE = _("Unable to delete the constraint.")
-SYNTAXERRORMSG = _("There is a syntax error in your %scode definition.")
-SERVERCONSTRAIN = _("%s Server constrain ")
-PREDEFINEDVARS = _(
-    "You specified a non predefined variable. The predefined variables are"
-    " self, datetime, dateutil, time, re, ValidationError and the ones"
-    " accessible through self, like self.env."
-)
-CONSTRAINEDLS = _(
-    "The field Constrained lists the fields that the server constrain will"
-    " check. It is a comma-separated list of field names, like name, size."
-)
-
-SAFE_EVAL_BASE["re"] = re
-SAFE_EVAL_BASE["ValidationError"] = ValidationError
-
-SAFE_EVAL_4FUNCTION = SAFE_EVAL_BASE
-SAFE_EVAL_4FUNCTION["api"] = api
-SAFE_EVAL_4FUNCTION["models"] = models
-SAFE_EVAL_4FUNCTION["fields"] = fields
-SAFE_EVAL_4FUNCTION["_"] = _
-
-
-def common_4constrains(el_self, code, message=SYNTAXERRORMSG):
-    """
-
-    :param el_self:
-    :param code:
-    :param message:
-    :return:
-    """
-
-    try:
-        safe_eval(code, SAFE_EVAL_4FUNCTION, {"self": el_self}, mode="exec")
-
-    except ValueError:
-        raise ValidationError(PREDEFINEDVARS)
-
-    except SyntaxError:
-        raise ValidationError(message)
-
-
-def add_constraint(cr, tablename, constraintname, definition):
-    """Add a constraint on the given table."""
-    query1 = 'ALTER TABLE "{}" ADD CONSTRAINT "{}" {}'.format(
-        tablename, constraintname, definition
-    )
-    query2 = 'COMMENT ON CONSTRAINT "{}" ON "{}" IS %s'.format(
-        constraintname, tablename
-    )
-    try:
-        with cr.savepoint():
-            cr.execute(query1)
-            cr.execute(query2, (definition,))
-            _logger.debug(
-                "Table %r: added constraint %r as %s",
-                tablename,
-                constraintname,
-                definition,
-            )
-            return True
-    except Exception:
-        msg = (
-            "Table %r: unable to add constraint %r!\nIf you want to have it,"
-            " you should update the records and execute manually:\n%s"
-        )
-        _logger.warning(msg, tablename, constraintname, query1, exc_info=True)
-        return False
-
-
-def drop_constraint(cr, tablename, constraintname):
-    """drop the given constraint."""
-    try:
-        with cr.savepoint():
-            cr.execute(
-                'ALTER TABLE "{}" DROP CONSTRAINT "{}"'.format(
-                    tablename, constraintname
-                )
-            )
-            _logger.debug(
-                "Table %r: dropped constraint %r", tablename, constraintname
-            )
-            return True
-    except ProgrammingError as proerror:
-        return proerror.pgerror.count("does not exist") or False
-
-    except Exception:
-        _logger.warning(
-            "Table %r: unable to drop constraint %r!",
-            tablename,
-            constraintname,
-        )
-        return False
-
-
-def sql_constraint(el_self, constraints):
-    cr = el_self._cr
-    foreign_key_re = re.compile(r"\s*foreign\s+key\b.*", re.I)
-
-    def process(key, definition):
-        conname = "%s_%s" % (el_self._table, key)
-        current_definition = tools.constraint_definition(
-            cr, el_self._table, conname
-        )
-        if not current_definition:
-            # constraint does not exists
-            return add_constraint(cr, el_self._table, conname, definition)
-        elif current_definition != definition:
-            # constraint exists but its definition may have changed
-            drop_constraint(cr, el_self._table, conname)
-            return add_constraint(cr, el_self._table, conname, definition)
-
-        else:
-            return True
-
-    result = False
-    for (sql_key, sql_definition, _) in constraints:
-        if foreign_key_re.match(sql_definition):
-            el_self.pool.post_init(process, sql_key, sql_definition)
-        else:
-            result = process(sql_key, sql_definition)
-
-        if not result:
-            break
-
-    return result
-
-
-class IrModelServerConstrain(models.Model):
-    _name = "ir.model.server_constrain"
-    _description = "Code Generator Model Server Constrains"
-    _rec_name = "constrained"
-
-    constrained = fields.Char(
-        string="Constrained",
-        help="Constrained fields, ej: name, age",
-        required=True,
-    )
-
-    @api.onchange("constrained")
-    @api.constrains("constrained")
-    def _check_constrained(self):
-        if self.constrained:
-            splitted = self.constrained.split(",")
-            if list(
-                filter(
-                    lambda e: e.strip()
-                    not in self.env[self.m2o_ir_model.model]._fields.keys(),
-                    splitted,
-                )
-            ):
-                raise ValidationError(CONSTRAINEDLS)
-
-    txt_code = fields.Text(
-        string="Code", help="Code to execute", required=True
-    )
-
-    @api.onchange("txt_code")
-    @api.constrains("txt_code")
-    def _check_txt_code(self):
-        if self.txt_code:
-            constrain_detail = SERVERCONSTRAIN % self.constrained
-            common_4constrains(
-                self.env[self.m2o_ir_model.model],
-                self.txt_code,
-                SYNTAXERRORMSG % constrain_detail,
-            )
-
-    m2o_ir_model = fields.Many2one(
-        comodel_name="ir.model",
-        string="Code generator Model",
-        help="Model that will hold this server constrain",
-        required=True,
-        domain=[("transient", "=", False)],
-        ondelete="cascade",
-    )
-
-
-class IrActionsReport(models.Model):
-    _inherit = "ir.actions.report"
-
-    m2o_model = fields.Many2one(
-        comodel_name="ir.model",
-        string="Code generator Model",
-        help="Model related with this report",
-        compute="_compute_m2os",
-        store=True,
-    )
-
-    m2o_template = fields.Many2one(
-        comodel_name="ir.ui.view",
-        string="Template",
-        help="Template related with this report",
-        compute="_compute_m2os",
-    )
-
-    @api.depends("model", "report_name")
-    def _compute_m2os(self):
-        for report in self:
-            searched = self.env["ir.model"].search(
-                [("model", "=", report.model.strip())]
-            )
-            if searched:
-                report.m2o_model = searched[0].id
-
-            stripped = report.report_name.strip()
-            splitted = stripped.split(".")
-            searched = self.env["ir.ui.view"].search(
-                [
-                    ("type", "=", "qweb"),
-                    ("name", "=", splitted[len(splitted) - 1]),
-                ]
-            )
-            if searched:
-                report.m2o_template = searched[0].id
 
 
 class IrModel(models.Model):
@@ -644,495 +418,3 @@ class IrModel(models.Model):
 #         self._run_safe_eval()
 #
 #         return result
-
-
-class IrModelDependency(models.Model):
-    _name = "code.generator.ir.model.dependency"
-    _description = "Code Generator ir model Dependency"
-
-    depend_id = fields.Many2one("ir.model", "Dependency", ondelete="cascade")
-    name = fields.Char(compute="compute_name")
-
-    ir_model_ids = fields.One2many(
-        comodel_name="ir.model",
-        inverse_name="inherit_model_ids",
-        string="Ir model",
-        help="Origin model with dependency",
-    )
-
-    @api.depends("depend_id")
-    def compute_name(self):
-        for rec in self:
-            if rec.depend_id:
-                rec.name = rec.depend_id.model
-
-
-class IrModelUpdatedFields(models.Model):
-    _name = "code.generator.ir.model.fields"
-    _description = "Code Generator Fields"
-
-    m2o_module = fields.Many2one(
-        "code.generator.module",
-        string="Module",
-        help="Module",
-        ondelete="cascade",
-    )
-
-    m2o_fields = fields.Many2one("ir.model.fields", string="Fields")
-
-    nomenclature_blacklist = fields.Boolean(
-        string="Ignore from nomenclature.", default=False
-    )
-
-    nomenclature_whitelist = fields.Boolean(
-        string="Force to nomenclature.", default=False
-    )
-
-    name = fields.Char(
-        string="Name",
-        help="Name of selected field.",
-        compute="_change_m2o_fields",
-    )
-
-    @api.onchange("m2o_fields")
-    def _change_m2o_fields(self):
-        for ir_field in self:
-            if ir_field.m2o_fields:
-                ir_field.name = ir_field.m2o_fields.name
-            else:
-                self.name = False
-
-
-FORCE_WIDGET_TYPES = [
-    ("barcode_handler", "Barcode handler"),
-    ("handle", "Handle"),
-    ("float_with_uom", "Float with uom"),
-    ("timesheet_uom", "Timesheet uom"),
-    ("radio", "Radio"),
-    ("priority", "Priority"),
-    ("mail_thread", "Mail thread"),
-    ("mail_activity", "Mail activity"),
-    ("mail_followers", "Mail followers"),
-    ("phone", "Phone"),
-    ("statinfo", "Statinfo"),
-    ("statusbar", "Statusbar"),
-    ("many2many", "Many2many"),
-    ("many2many_tags", "Many2many tags"),
-    ("many2many_tags_email", "Many2many tags email"),
-    ("many2many_checkboxes", "Many2many checkboxes"),
-    ("many2many_binary", "Many2many binary"),
-    ("monetary", "Monetary"),
-    ("selection", "Selection"),
-    ("url", "Url"),
-    ("boolean_button", "Boolean button"),
-    ("boolean_toggle", "Boolean toggle"),
-    ("toggle_button", "Toggle button"),
-    ("state_selection", "State selection"),
-    ("kanban_state_selection", "Kanban state selection"),
-    ("kanban_activity", "Kanban activity"),
-    ("tier_validation", "Tier validation"),
-    ("binary_size", "Binary size"),
-    ("binary_preview", "Binary preview"),
-    ("char_domain", "Char domain"),
-    ("domain", "Domain"),
-    ("file_actions", "File actions"),
-    ("color", "Color"),
-    ("copy_binary", "Copy binary"),
-    ("share_char", "Share char"),
-    ("share_text", "Share text"),
-    ("share_binary", "Share binary"),
-    ("selection_badge", "Selection badge"),
-    ("link_button", "Link button"),
-    ("image", "Image"),
-    ("contact", "Contact"),
-    ("float_time", "Float time"),
-    ("image-url", "Image-url"),
-    ("html", "Html"),
-    ("email", "Email"),
-    ("website_button", "Website button"),
-    ("one2many", "One2many"),
-    ("one2many_list", "One2many list"),
-    ("gauge", "Gauge"),
-    ("label_selection", "Label selection"),
-    ("percentpie", "Percentpie"),
-    ("progressbar", "Progressbar"),
-    ("mrp_time_counter", "Mrp time counter"),
-    ("qty_available", "Qty available"),
-    ("ace", "Ace"),
-    ("pdf_viewer", "Pdf viewer"),
-    ("path_names", "Path names"),
-    ("path_json", "Path json"),
-    ("date", "Date"),
-    ("color_index", "Color index"),
-    ("google_partner_address", "Google partner address"),
-    ("google_marker_picker", "Google marker picker"),
-    ("spread_line_widget", "Spread line widget"),
-    ("geo_edit_map", "Geo edit map"),
-    ("dynamic_dropdown", "Dynamic dropdown"),
-    ("section_and_note_one2many", "Section and note one2many"),
-    ("section_and_note_text", "Section and note text"),
-    ("reference", "Reference"),
-    ("x2many_2d_matrix", "X2many 2d matrix"),
-    ("numeric_step", "Numeric step"),
-    ("BVEEditor", "Bveeditor"),
-    ("er_diagram_image", "Er diagram image"),
-    ("u2f_scan", "U2f scan"),
-    ("password", "Password"),
-    ("open_tab", "Open tab"),
-    ("signature", "Signature"),
-    ("upgrade_boolean", "Upgrade boolean"),
-    ("many2manyattendee", "Many2manyattendee"),
-    ("res_partner_many2one", "Res partner many2one"),
-    ("hr_org_chart", "Hr org chart"),
-    ("CopyClipboardText", "Copyclipboardtext"),
-    ("CopyClipboardChar", "Copyclipboardchar"),
-    ("bullet_state", "Bullet state"),
-    ("pad", "Pad"),
-    ("field_partner_autocomplete", "Field partner autocomplete"),
-    ("html_frame", "Html frame"),
-    ("task_workflow", "Task workflow"),
-    ("document_page_reference", "Document page reference"),
-    ("mis_report_widget", "Mis report widget"),
-    ("kpi", "Kpi"),
-    ("action_barcode_handler", "Action barcode handler"),
-    ("mail_failed_message", "Mail failed message"),
-    ("mermaid", "Mermaid"),
-    ("payment", "Payment"),
-    ("previous_order", "Previous order"),
-]
-
-
-class IrModelFields(models.Model):
-    _inherit = "ir.model.fields"
-
-    force_widget = fields.Selection(
-        FORCE_WIDGET_TYPES,
-        string="Force widget",
-        help="Use this widget for this field when create views.",
-    )
-
-    default = fields.Char(string="Default value")
-    is_show_whitelist_list_view = fields.Boolean(
-        string="Show in whitelist list view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View list only."
-        ),
-    )
-    is_hide_blacklist_list_view = fields.Boolean(
-        string="Hide in blacklist list view",
-        help="Hide from view when field is blacklisted. View list only.",
-    )
-    is_show_whitelist_form_view = fields.Boolean(
-        string="Show in whitelist form view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View form only."
-        ),
-    )
-    is_hide_blacklist_form_view = fields.Boolean(
-        string="Hide in blacklist form view",
-        help="Hide from view when field is blacklisted. View form only.",
-    )
-    is_show_whitelist_kanban_view = fields.Boolean(
-        string="Show in whitelist kanban view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View kanban only."
-        ),
-    )
-    is_hide_blacklist_kanban_view = fields.Boolean(
-        string="Hide in blacklist kanban view",
-        help="Hide from view when field is blacklisted. View kanban only.",
-    )
-    is_show_whitelist_search_view = fields.Boolean(
-        string="Show in whitelist search view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View search only."
-        ),
-    )
-    is_hide_blacklist_search_view = fields.Boolean(
-        string="Hide in blacklist search view",
-        help="Hide from view when field is blacklisted. View search only.",
-    )
-    is_show_whitelist_pivot_view = fields.Boolean(
-        string="Show in whitelist pivot view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View pivot only."
-        ),
-    )
-    is_hide_blacklist_pivot_view = fields.Boolean(
-        string="Hide in blacklist pivot view",
-        help="Hide from view when field is blacklisted. View pivot only.",
-    )
-    is_show_whitelist_graph_view = fields.Boolean(
-        string="Show in whitelist graph view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View graph only."
-        ),
-    )
-    is_hide_blacklist_graph_view = fields.Boolean(
-        string="Hide in blacklist graph view",
-        help="Hide from view when field is blacklisted. View graph only.",
-    )
-    is_date_start_view = fields.Boolean(
-        string="Show start date view",
-        help="View timeline only, start field.",
-    )
-    is_date_end_view = fields.Boolean(
-        string="Show end date view",
-        help="View timeline only, end field.",
-    )
-    is_show_whitelist_calendar_view = fields.Boolean(
-        string="Show in whitelist calendar view",
-        help=(
-            "If a field in model is in whitelist, all is not will be hide. "
-            "View calendar only."
-        ),
-    )
-    is_hide_blacklist_calendar_view = fields.Boolean(
-        string="Hide in blacklist calendar view",
-        help="Hide from view when field is blacklisted. View calendar only.",
-    )
-    # This is used to choose order to show field in model
-    code_generator_sequence = fields.Integer(
-        string="Sequence Code Generator",
-        help="Sequence to write this field from Code Generator.",
-    )
-
-    # TODO remove code_generator_tree_view_sequence and code_generator_search_sequence
-    # TODO wrong architecture, separate view order from model
-    # TODO This was a work around
-    # TODO or maybe it's useful in first iteration of code generator, remove this later when A USE C GENERATE B
-    code_generator_tree_view_sequence = fields.Integer(
-        string="Tree view sequence",
-        help="Sequence to write this field in tree view from Code Generator.",
-        default=-1,
-    )
-
-    code_generator_form_simple_view_sequence = fields.Integer(
-        string="Form simple view sequence",
-        help=(
-            "Sequence to write this field in form simple view from Code"
-            " Generator."
-        ),
-        default=-1,
-    )
-
-    code_generator_search_sequence = fields.Integer(
-        string="Search sequence",
-        help="Sequence to write this field in search from Code Generator.",
-        default=-1,
-    )
-
-    code_generator_kanban_view_sequence = fields.Integer(
-        string="Kanban view sequence",
-        help=(
-            "Sequence to write this field in kanban view from Code Generator."
-        ),
-        default=-1,
-    )
-
-    code_generator_calendar_view_sequence = fields.Integer(
-        string="calendar view sequence",
-        help=(
-            "Sequence to write this field in calendar view from Code"
-            " Generator."
-        ),
-        default=-1,
-    )
-
-    code_generator_search_view_sequence = fields.Integer(
-        string="search view sequence",
-        help=(
-            "Sequence to write this field in search view from Code Generator."
-        ),
-        default=-1,
-    )
-
-    code_generator_pivot_view_sequence = fields.Integer(
-        string="pivot view sequence",
-        help="Sequence to write this field in pivot view from Code Generator.",
-        default=-1,
-    )
-
-    code_generator_graph_view_sequence = fields.Integer(
-        string="graph view sequence",
-        help="Sequence to write this field in graph view from Code Generator.",
-        default=-1,
-    )
-
-    code_generator_compute = fields.Char(
-        string="Compute Code Generator",
-        help="Compute method to code_generator_writer.",
-    )
-
-    ignore_on_code_generator_writer = fields.Boolean(
-        help="Enable this to ignore it when write code."
-    )
-
-    @api.constrains("name", "state")
-    def _check_name(self):
-        for field in self:
-            if field.state == "manual":
-                if (
-                    not field.model_id.m2o_module
-                    and not field.name.startswith("x_")
-                ):
-                    raise ValidationError(
-                        _(
-                            "Custom fields must have a name that starts with"
-                            " 'x_' !"
-                        )
-                    )
-            try:
-                models.check_pg_name(field.name)
-            except ValidationError:
-                msg = _(
-                    "Field names can only contain characters, digits and"
-                    " underscores (up to 63)."
-                )
-                raise ValidationError(msg)
-
-    @api.model
-    def create(self, vals):
-        if "model_id" in vals:
-            model_data = self.env["ir.model"].browse(vals["model_id"])
-            vals["model"] = model_data.model
-        if vals.get("ttype") == "selection":
-            if not vals.get("selection"):
-                raise UserError(
-                    _(
-                        "For selection fields, the Selection Options must be"
-                        " given!"
-                    )
-                )
-            self._check_selection(vals["selection"])
-
-        res = super(models.Model, self).create(vals)
-
-        if vals.get("state", "manual") == "manual":
-
-            check_relation = True
-            if vals.get("relation") and vals.get("model_id"):
-                check_relation = not model_data.m2o_module
-
-            if (
-                vals.get("relation")
-                and not self.env["ir.model"].search(
-                    [("model", "=", vals["relation"])]
-                )
-                and check_relation
-            ):
-                raise UserError(
-                    _("Model %s does not exist!") % vals["relation"]
-                )
-
-            if vals.get("ttype") == "one2many":
-                # TODO check relation exist, but some times, it's created later to respect many2one order
-                # if not self.env[""].search(
-                #     [
-                #         ("model_id", "=", vals["relation"]),
-                #         ("name", "=", vals["relation_field"]),
-                #         ("ttype", "=", "many2one"),
-                #     ]
-                # ):
-                #     raise UserError(
-                #         _("Many2one %s on model %s does not exist!")
-                #         % (vals["relation_field"], vals["relation"])
-                #     )
-                pass
-
-            self.clear_caches()  # for _existing_field_data()
-
-            if vals["model"] in self.pool:
-                # setup models; this re-initializes model in registry
-                self.pool.setup_models(self._cr)
-                # update database schema of model and its descendant models
-                descendants = self.pool.descendants(
-                    [vals["model"]], "_inherits"
-                )
-                self.pool.init_models(
-                    self._cr,
-                    descendants,
-                    dict(self._context, update_custom_fields=True),
-                )
-
-        return res
-
-
-class IrModelConstraint(models.Model):
-    _inherit = "ir.model.constraint"
-
-    code_generator_id = fields.Many2one(
-        comodel_name="code.generator.module",
-        string="Code Generator",
-        ondelete="cascade",
-    )
-
-    module = fields.Many2one(
-        default=lambda self: self.env["ir.module.module"]
-        .search([("name", "=", "base")])[0]
-        .id
-    )
-
-    model_state = fields.Selection(related="model.state")
-
-    message = fields.Char(string="Message", help="Message")
-
-    @api.model_create_multi
-    def create(self, vals_list):
-        imcs = super(IrModelConstraint, self).create(vals_list)
-
-        for imc in imcs:
-            m_target = self.env[imc.model.model]
-            t_constrain = (imc.name, imc.definition, imc.message)
-            if not sql_constraint(
-                m_target, m_target._sql_constraints + [t_constrain]
-            ):
-                raise ValidationError(CONSCREATEUNABLE)
-
-        return imcs
-
-    @api.multi
-    def write(self, vals):
-
-        if "definition" in vals:
-            for imc in self:
-                tablename = self.env[imc.model.model]._table
-                if not drop_constraint(
-                    self._cr, tablename, "%s_%s" % (tablename, imc.name)
-                ):
-                    raise ValidationError(CONSDELETECREATEUNABLE)
-
-        result = super(IrModelConstraint, self).write(vals)
-
-        for imc in self:
-            m_target = self.env[imc.model.model]
-            t_constrain = (imc.name, imc.definition, imc.message)
-            if not sql_constraint(m_target, [t_constrain]):
-                raise ValidationError(CONSMODIFYUNABLE)
-
-        return result
-
-    @api.multi
-    def unlink(self):
-
-        for imc in self:
-            try:
-                tablename = self.env[imc.model.model]._table
-                if not drop_constraint(
-                    self._cr, tablename, "%s_%s" % (tablename, imc.name)
-                ):
-                    raise ValidationError(CONSDELETEUNABLE)
-
-            except MissingError:
-                _logger.warning(
-                    "The registry entry associated with %s no longer exists"
-                    % imc.model.model
-                )
-
-        return super(IrModelConstraint, self).unlink()
