+++ code_generator/models/ir_module_module.py
@@ -1,0 +1,9 @@
+from odoo import models, fields, api, modules, tools
+
+
+class IrModuleModule(models.Model):
+    _inherit = "ir.module.module"
+
+    header_manifest = fields.Text(
+        string="Header", help="Header comment in __manifest__.py file."
+    )
